requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
multidict==6.1.0
propcache==0.2.0
python-telegram-bot==21.6
sniffio==1.3.1
yarl==1.14.0
